- Fix some memory leaks - "Better" debouncing. Parameter is currently mis-named, but in the future I plan on learning how to use the timers to make it actually MS-based (before final release) - Display of rules

+++ rule.h
@@ -5,8 +5,8 @@
 #include <stdlib.h>
 
 typedef struct Rule {
-    uint8_t Live;
-    uint8_t Born;
+    int Live;
+    int Born;
     char* Name;
 } Rule;
